feat(demo): add sleep and wakeup time gifs
This commit adds new gif files for demonstrating the functionality of the app. It includes gifs for figuring out when to wake up, when to go to sleep, and when to wake up if you go to sleep now. These gifs enhance the visual representation of the app's features in the README file.